This is a tentative workaround for a problem I ran into while debugging a threading problem.
When a pure python "daemon" thread uses
performSelectorOnMainThread:... and to call a
method on the main thread and the called method
raises an exception we might end up in the
situation where we want to use the PyGILState
API's when the machinery for that has shut down
(issue 1402 in python's bugtracker). This causes
a hard crash (abort() is called by the Python
core).

By calling objc.recycleAutoreleasePool in an
atexit handler we avoid running in that issue.

This is a tentative fix because I haven't worked
out yet if this might cause other problems.