My own version of Leopard's metadata scanning tools.
This is mostly for historical interest, I've used these
tools to shape my thoughts about the bridgesupport tools
and provide feedback to the bridgesupport tools.

I am planning further development though, mostly to
experiment with solutions to problems I'm finding in the
metadata system. The most important of which are
dealing with the C++ type bool (cannot be represented
in metadata at the moment) and unicode strings (no nice
representation at the moment).