Add an new way for specifying protocol conformance in Python 3.x
It is now possible to write a class like this:

   class MyObject (NSObject, protocols=[SomeProtocol]):
      pass

The old way of using protocols as mix-in classes still works::

   class MyObject (NSObject, SomeProtocol):
      pass

The new syntax only works for Python 3.x (it is a syntax error
in Python 2.x).


The new was added because the older method does not work with the current
python 3.3 repository, and because it better documents what the
code intends to accomplish.