Change calculation of default method signature to be more sane
In previous releases the default Objective-C method signature was
calculated from the Python method signarure. In this version the
signature is calculated from the implied Objective-C selector.

This makes it possible to use varargs in the defition of selectors:

   def someMethod_arg_(self, *args): ...

This in turn makes it a lot easier to use generic decorators in
Cocoa classes.

This is a backward incompabible change, a number of patterns will
cause an exception at class definition time instead of generating
an Objective-C method signature that doesn't match the implied
prototype of the Objective-C selector. An example of that:

   def myMethod_arg_(self, a): ...

More importantly, this will cause problems when mixing plain python
methods and Objective-C selectors in a class definition. The workaround
is to add ``objc.python_method`` as a decorator for methods that are
only called from Python:

   @objc.python_method
   def myhelper(self, a, b): ...

Fixes #143