Fix crash due to partialy contstructed class in ObjC runtime
Creating a class that claims to conform to a protocol but
doesn't raises an exception. In some circumstances the
Objective-C part of the class was created an registered with
the Objective-C runtime and could be accessed later on. Actually
using the class would then cause a hard crash.

This commit also changes the way in which pyobjc-core keeps
class definitions alive: instead of artificially increasing
the refcount of the class all classes are added to a
(private) list.