Remove PyObjCTools.TestSupport.filterWarnings
The function was not necessary and accessed internals of the
warnings module.

This also fixes a remaining error in test_structs when running
the entire test suite.

Also don't emit a DeprecationWarning in the tp_as_mapping functions
in struct-wrapper.m, those will be emitted in the as_sequence
functions.