Move documentation to a central location in the repository.
This also removed the need to have a separate pyobjc-website repository
with scripts to collect files into the right locations.

TODO: Integrate examples into the new structure.

The major reason for doing this is to make it possible to use
readthedocs to host the documentation, which will remove some
needless manual labor from updating the website.

A nice side effect is that this should make it easier to work
on the documentation.