#278: Surpress -Wunguarded-availability when compiling _inlines extensions
These modules contain a static array of function pointers to inline
functions, which ensures that those functions get included into the
extension module.

The warning -Wunguarded-availability causes false
positives for these definitions, these functions will definitely be
available at runtime because the function is included in the binary
and not referenced from a framework.