pyobjc_setup.py: sdk detection fixes
Detect correct SDK when an -isysroot flag is given in CFLAGS in the
environment, overriding any that may be in sysconfig. Support alternate
form of -isysroot where there is no space between the flag and the
associated path. Don't try to extract a version from the filename of
'MacOSX.sdk'; instead get the version from its SDKSettings.plist.