Introduce "simple" callers
For simple enough method/function signatures we can do
away with a lot of the complexity of calling, and in particularly
the overhead of dynamic memory allocations.

This gives a clear performance improvement for calling
selectors, and especially functions.

I expect to change the definition of what's a "simple"
signature over time (not at least because I'm not happy
with the code duplication).