Switch the creation of opaque pointer wrappers to PyType_FromSpec
The previous code was poking into CPython internals while there
is a nicer public API availalble.

I don't like the lack of type checking in the PyType_FromSpec
interface, but the new code is cleaner than previously.

Also add a tp_traverse implementation which is needed for
Python 3.9 or later (see #364)