Improve test coverage for opaque-pointer.m #393
This also makes opaque-pointer support the GC protocol,
needed because of the strong reference to the type
in Python 3.9 or later.

I'm not 100% happy with the way I test the traverse
function, but I haven't found a way yet to trigger
a call to the slot in a normal way.