Test cleanup
This removes some functions and methods from PyObjCTools.TestSupport
that are either no longer relevant ("onlyOnPython2", "onlyOn32Bit"),
or have better replacements in unittest ("onlyIf" -> "skipUnless").

Tests and documentation are updated to match.

The changelog entry is fairly minimal, but I expect that the PyObjC
testsuite is the only user of this module.