Improve test coverage for _lazyloader.py
This also fixes a bug that should not affect anyone:
fetching __all__ could raise AttributeError when trying
to load a magic indirect constant that isn't present.

This brings coverage of the Python code in pyobjc-core
back to 100%.  With one snag: "make coverage" still reports
99% coverage for _lazyloader.py due to 2 incompletely covered
branches, but running just test_lazyloader.py shows 100% coverage
for that file.