Move code dealing with libffi closure API to seperate file
Due to having to deal with a number of API variants
for compilation and running on multiple macOS versions
the closure related code was hard to read and maintain.

The preprocessor logic is now centralized in a single file.

This also removes closure_pool.h from pyobjc.h because
it is only used by libffi_extra.m.  Both closure_pool.m
and libffi_extra.m don't include pyobjc.h because they
are standalone.

Fixes #317