Fix edge case in default implementation of ``-copy``
PyObjC introduces a default implemetentation of -copy
for the first Python generation of subclasses of
an ObjC class that has a -copy selector. This default
implementation adjust Python refcounts and copies
``__dict__``.

The implementation did not consider that -copy may
return an instance of a different class and would
crash when that happens.