Add some helpers for test #393
This adds a new class to PyObjCTest.objectint that
can (by default) not be represented in Python, and
will be used later to test some error paths.

Also add helpers to make it easier to add tests with
this class for arrays, dictionaries and sets easier.

The actual tests will be added later.