Use the correct setuptools install command
Trying to install pyobjc-framework-ScrenCaptureKit on a 12.2 system
should give an error message that explains that this package requires
12.3:

   error: This distribution is only supported on MacOSX >= 12.3

This didn't work when the "wheel" package was not installed. Pip
than uses a different code path using
``--single-version-externally-managed`` and that didn't work due
to pyobjc_setup.py using distutils.command.install instead of
setuptools.command.install.

Issue #455