Fix handling of _C_OUT argument when the base type is a CoreFoundation type
This is related to #456, and avoids hitting an internal assertion
error with invalid metadata for a function/method. The assertion
error indicates a programming error in the bridge itself, and those
assertions are something that I'll disable for production builds
in a future version (which would have resulted in memory corruption
in this case).