Error out when building wheels fails
The build script only warned when the wheel for
a subproject cannot be build. That's a leftover from
a period when wheels needed to be build on two systems
to support some frameworks that were no longer in the
latest SDK.

Currently all wheels can be build with the latest SDK,
and warning can hide build problems.

Issue #468