Fix ``build_ext`` for bindings without C extension
The custom ``build_ext`` command in PyObjC's framework
bindings assumed that it will only be invoked when the
binding contains a C extension and inserts an unnecessary
build dependency on pyobjc-core when the extension doesn't
contain one.

This changeset fixes this problem by changing the test
that checks if a dependency on pyobjc-core is necesary.

Fixes #468.