Issue #94: Add objc.simd module
This module defines a number of Python types that
implement the interface of a subset of SIMD types
defined in <simd/simd.h>.

Only those types used in frameworks are provided.

The implementation and tests are still work in
progress, but should be good enough to start work
on supporting these in the core bridge.

The implementation is in Python for easy of implementation,
can be replaced by a C implementation later when that is
needed for performance.