Issue #492: Add metadata for "_Tundra" class clones in AVFoundation
As described in #492 there is a second class hierarchy in AVFoundation
uses "_Tundra" in the class name.  For all metadata updates for regular
classes also register that metadata for the "_Tundra" clone.

This will fix the issue in PyObjC 9, there will be a separate fix
for the current release branch.