Switch to using imp_implementationWithBlock to create IMPs
This switches to using imp_implementationWithBlock for creating
IMPs with a manual binding (the FFI path is unchanged), primarily
because this makes it possible to provide IMPs for selectors
with vector arguments/return values.

Also updated the helpers-vectors.m code (and generator script) to
use this functionality. I expect more changes to these helpers
when I start on proper tests for vector types in method calls.