Issue #306: objc.objc_class.__setattr__ now uses Python transformer
The implementation of objc_class.__setattr__ now calls
the attribute transformer in objc._transform instead of
the equivalent C code.

This also found an incorrect edge case in determining
if the bytecode of a function indicates that the function
returns a value.

This is the first step in completely replacing the C logic
by the code in objc._transform, class creation still uses
the C logic.