Issue #306: Slightly refactor class-builds.m
This moves the logic that will be moved to Python to a seperate
function, and exposes that function to Python for testing.

Also uses that newly exposed function to run the tests for the
new implementation on  the old one, and disables a small number
of tests that are known to show different behaviour.