Switch to using the Python implmenentation
The older implementation is not yet removed for easier
comparison of the Python and C implementation during debugging.

A small number of tests required changes due to semantic
changes in a number of edge cases.

Also introduces a failing test related to hidden selectors,
which already is a problem in current releases found while
working on the Python implementation.

Almost there...

Issue #306