Move protocol conformance checking into Python
This changeset moves the logic for checking if a
newly defined class conforms to the classes it claims
to conform to into Python.

To make this possible this also adds ``instanceMethods``
and ``classMethods`` methods to ``objc.informal_protocols``,
with an interface mirroring that of ``objc.formal_protocol``.

Next up is moving ``objc.informal_protocol`` to Python.