Reimplement the forwardInvocation: method of Python subclass of NSObject
The old implementation tried to reimplement the logic
in the method_stub function in libffi_support, the new one
just uses libffi to call that stub. That simplifies the implemention
and ensures the two code paths are consistent.

Issue #523 proposes dropping the implementation altogether,
reimplementing it reduces the risk of breaking user code while still
reducing implementation complexity.

Fixes #523.