Always introduce an intermediate class for a first-gen Python class
Issue #522

PyObjC already introduced an intermediate class when necessary,
this is now done unconditionally. This simplifies the code, but
also makes sure that adding methods like ``copyWithZone:`` after
the class is created works correctly.

This also fixes a bug in objc.listInstanceVariables(): That function
would fail if one of the classes in the class hierarchy didn't have
instance variables at all.