Fix edge case in Py_buffer usage in the NSCoder helpers
Issue #363.

The code assumed that the Py_buffer.buf value is NUL-terminated,
which isn't a valid assumption (inherited from previous code that
used the 'y' specifier with PyArg_Parse*). This can lead to reading
beyond the end of the buffer (e.g. when the type string passed to
these methods is incomplete).

Fixed by adding a function that validates a type string while keeping
track of the length.