Issue #539: Bad metadata for IOSurfaceCreate
The function was marked having an increased refcount for
both CoreFoundation and Objective-C, instead of just for
one of them. That results in releases the reference during
creation of the Python proxy for the return value, which later
on results in a crash.

Adding tests found the same issue in CoreServices and Metal
bindings as well.