Issue #552: Fix build and test errors on macOS 10.11
This fixes some build errors on macOS 10.11 and also fixes a number
of issues with the test suite that resulted in test failures when
doing a full test run (with python 3.11).

There are a number of edge cases where test failures are suppressed
at the moment (such as handling proxying 2**64 to an NSNumber), those
will be fixed later and shouldn't affect user code.