Switch away from ObjCLazyModule for all framework bindings
This rewrites the __init__.py of all framework bindings to
switch away from ObjCLazyModule to objc.createFrameworkDirAndGetattr.

A side effect of this is that those init files are now a
lot cleaner than before (e.g. cruft removal).

Also adds a test for validating the bundle identifier of
frameworks, and fixes those identifiers in a couple of bindings.

Closes #295 and closes #561.