Fix incompatibility with macOS 14 beta
The beta introduces some clases with non-identifier characters
in their name (in particular angled brackets). Switch the lazy loader
to test for valid identifiers, instead of blacklisting dots in names.

This is a backport of a similar fix in the 10.0 branch. I haven't
cherrypicked that changeset because there are other changes to the lazy
loader in that branch that aren't suitable for backporting.

Fixes #563