Fix crash when the return value in a type encoding returned from the ObjC runtime contains a class name
For some (private) classes and methods on macOS 14 the return value can
contain an annation with the class name:

    @"FBSDisplayConfiguration"16@0:8

That is, the return type is ``id`` with the added information that
this is an instance of ``FBSDisplayConfiguration``. The annotation was
already handled for argument types, but not for return types.

This changeset ignores the annotation for return types as well (instead
of causing a hard crash).