Implement NS_UNAVAILABLE support through the metadata system
There's still an API to support __new__ keyword support, but
actually marking the selector unavailable is done through the
metadata system.

Also add tests for the NS_UNAVAILABLE support code.

A minor disadvantage of this is that subclasses cannot mark a
method as available again, but AFAIK that's not used in Apple's
SDKs and can be added later.