Support generic new functionality for NSArray classes
These classes already had a __new__ method, restructure
the code to support both generic new feature and the older
Python compatible interface.

This breaks using a keyword argument named "sequence", support
was dropped to get an easier to explain interface. This is compatible
with the interface for list and tuple because don't accept keyword
arguments at all.