Drop using PySequence_Fast in pyobjc-core
PySequence_Fast has semantics is not ideal in general, and
more so in free threaded mode: when the argument is a python
list it is returned as is, which means the borrowed references
returned from ..._GET_ITEM are problematic due to concurrent
modification of the list.

Switch to a new ``PyObjCSequence_Tuple`` API that is a version
of ``PySequence_Tuple`` with a message that gives more context
to th user. The error raised by ``PySequene_Tuple`` is used as the
``__cause__`` of the exception raised by the new API.

Issue #608