Fix a race condition in creating Python proxies for ObjC classes
Issue #619

PyObjC calls back into the Python interpreter while creating a class
proxy object and that can result in giving up the GIL for a short time.

This gives a small window where another thread can create a proxy
for the same class, and that in the end results in deallocating the
proxy for one of the two Python classes.

This PR fixes the issue by checking for the race and cleaning up the
unnecessary second proxy class without triggering a warning for users.