Issue #608: Turn accessor macros in objc-object.h into functions
This is a first step in making objc-object.m safe to use
in a free-threaded build.

The change to method-accessor.m fixes a real bug: a PyObjCClass
instance was used as a PyObjCObject one. This happens to work
due to both having the Objective-C value as their first field,
but is undefined behaviour in C.