Issue #608: Free-threading support in objc-object.m
objc-object is now almost ready for free-threading, with the
exception of usage of '->flags': I haven't audited the code
yet to see if that's safe or needs more work (most flags are static,
but not all).

The rest of the code is not 100% safe with free-threading, but the
safeness issues should also be present in when using the GIL: calling
an 'init' method on an UNINITIALIZED value in two threads at once
has a race condition where one of the threads uses a no longer valid
value. That's close to impossible to guard against.