Rework 'PyObjC_SELToPythonName'
This function was passed a stack buffer that was for the
most part converted to a string, return a Python string
directly and deal with the fallout.

The primary reason to do this is that I was using inconsitent
buffer sizes, and it turns out that selector names get very
long for some (private) APIs on macOS 15.