Add metadata introspection field for initializers
The result of ``__metadata__()`` can now contain an ``initializer``
field with value ``True`` when the selector is an init method
according to the ObjC naming convention.

This field cannot be changed through the metadata system.

The same dictionary can also contain ``free_result`` when the callable
returns a C pointer that needs to be free-ed using free(3) (which is
automaticly done by the bridge). This flag can already be set through
the metadata system, but was not introspectable.

Both flags can be tested for using methods of ``PyObjCTools.TestSupport.TestCase``

The ``initializer`` flag is infrastructure for future cleanup
of the bridge.