Update handling of alloc and init methods
This changeset fixes a longstanding design issue with PyObjC's bridge.

Before this changeset PyObjC used an "uninitialized" state for objects
that are the result of an alloc method, and clears that state after
calling an init method, with special handling when the init method does
not return *self*.

That's not in line with the semantic model for handling reference counts
in Objective-C.  This changeset fixes that (using Python lingo for the
types of references):

1. The result of a method in the alloc family is a new reference
   (unchanged in the brigdge)

2. A method in the init family steals a reference to self and returns
   a new reference.

   The bridge implements this by retaining self before calling the
   init method and handling the result as an 'already_retained' value.

   This brings us in line with the ARC documentation.

This changeset took a long time to flesh out because the change caused
us to hit another problem: For (some) classes implemented in Swift it
is not safe to use ``CFRelease`` to release the +1 reference returned
from an alloc method.

Because of that PyObjCObject_Type now uses normal reference count
updates instead of CFRetain and CFRelease. The CF calls were introduced
to support Objective-C Garbage Collection which we haven't supported in
a very long time.