Use old setuptools for c-coverage build
Recent versions of setuptools change how 'python setup.py develop'
works, and that breaks the Makefile entry for C-level coverage
testing.

As a workaround install an older version of setuptools from before
this change.

That's not a corrct fix, but at least makes it possible to do
C-level coverage tests.

See also #497