Move actual object deallocation inside the proxy-registery mutex
Before this changeset object_dealloc and fetching a value
from the proxy registry could race. This can result in a
use-after-free.

The new free-threading test script found this issue, but
this changeset does not fully fix the issue here.

I'm not entirely happy with the code patter, in particular
with testing that ``Py_REFCNT(obj) == 0``, but seems to
work...