thread-safety in updating class slots
Only call type.__setattr__ when the new attribute
is different from the old one in the free-threaded
build.

In 3.13 and 3.14 this protects against a thread
safety issue in the implementation of the type
class (in the sense of making it more unlikely to
hit the issue)

In 3.15 the thread safety issue is fixed, but this
makes updating slots more expensive (hence something
to avoid when the value doesn't actually change).

Issue #608