Remove old metadata scan files
Framework bindings contain scans of Apple framework
headers, which are compiled into metadata files for
definitions that cannot be extracted from the Objective-C
runtime.

This moves us to having 1 metadata scan for every framework,
although still in x86_64 and arm64 due to having a (very) small
number of differenes.

Legacy definitions are now maintained as manual updates.

This should make it easier to evolve the tooling, which
is needed for a couple of upcoming features.