Address some of the XXX comments in pyobjc-core
The only user-visible change is that creating a copy
of a number from Objective-C through ``-[OC_PythonNumber
copyWithZone:]`` now actually copies when the value
is not a builtin number. For the builtin numbers we know
the value is immutable and doesn't need to be copied.

The other changes are code cleanup.